requirements.txt
fastapi
uvicorn
torch
transformers
pillow
pydantic
python-multipart
accelerate